requirements.txt
uvicorn
librosa
numpy
matplotlib
torch
transformers
pydub
Pillow
openai-whisper
